python/requirements.txt
torch
transformers
sentencepiece
protobuf
python-docx
windnd
